feat:command split
