+++ src/Services/Masa.Auth.Service.Admin/appsettings.Develop.json
@@ -21,6 +21,6 @@
   "PmClient": {
     "Url": "https://pm-service-staging.masastack.com/"
   },
-  "WebUrl": "https://auth-develop.masastack.com/",
+  "WebUrl": "https://auth-develop.masastack.com",
   "IdentityServerUrl": "https://sso-develop.masastack.com/"
 }
