+++ src/Services/Masa.Auth.Service.Admin/appsettings.Staging.json
@@ -21,7 +21,7 @@
   "PmClient": {
     "Url": "https://pm-service-staging.masastack.com/"
   },
-  "WebUrl": "https://auth-staging.masastack.com/",
+  "WebUrl": "https://auth-staging.masastack.com",
   "IdentityServerUrl": "http://sso-staging.masastack.com/",
   "AutoComplete": [
     {
